Revert "Revert "De test.py se renombró a main.py""
This reverts commit b9f040c85f41bad0362d50845217699b80218b47.
