+++ src/manejadorDeSenales.py
@@ -1,0 +1,7 @@
+from senalDiscreta import SenalDiscreta
+
+def obtenerSenalDiscretaDesdeGrabacion():
+    print()
+
+def crearAudioDesdeSenalDiscreta():
+    print()
